Add class=pond to the water layer (#1456)
This PR proposes to add a new class, "pond" to the water layer, which allows for a style to de-emphasize minor bodies of water (tagged `water=pond,basin,wastewater`) when compared to more significant objects tagged reservoir/lake.  This PR retains the default behavior of classifying a water body as a "lake" if more specific tagging is not added.