Remove swimming pool from water_name (#1748)
Presently, swimming pool labels are rendered as lake points in the `water_name` layer.  This produces absurd results as follows:

In #1321, we made a separate swimming pool class so that they could be differentiated from lakes.  However, we did not adjust the labels.

In fact, we now have two nodes - a `water_name` node and a `poi` node:

To correct this, this PR simply removes swimming pool labels from the `water_name` layer, as labeling a swimming pool the same way as a lake is clearly wrong.  If there is a need to put labels on swimming pools, styles can continue to use the nodes available in the `poi` layer.